requirements.txt
Jinja2==3.1.4
PyYAML==6.0.2
pylint==3.2.6
black==24.8.0
